saas/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
confluent-kafka>=2.3.0
python-dotenv>=1.0.0
pydantic>=2.0.0
